requirements.txt
langchain
langchain-core
langchain_community
langchain-google-genai
google-generativeai
python-dotenv
langgraph
pydantic
langchain-openai
